+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Note: Prior to version 0.3.4.0, this library was named
 `small-bytearray-builder` is now just a compatibility shim
 to ease the migration process.
 
+## 0.3.15.0 -- 2023-??-??
+
+* Add `Data.Bytes.Builder.Unsafe.pasteUtf8TextJson#` for users who need
+  to perform JSON string encoding without using a builder.
+
 ## 0.3.14.0 -- 2023-07-20
 
 * Add `runOntoLength`.